Lager interfacer, snur dependencies, bereder grunnen til å kunne flytte servicene inn i domene
