+++ app/src/main/kotlin/no/nav/tiltakspenger/vedtak/ApplicationBuilder.kt
@@ -4,16 +4,20 @@ import mu.KotlinLogging
 import no.nav.helse.rapids_rivers.RapidApplication
 import no.nav.helse.rapids_rivers.RapidsConnection
 import no.nav.tiltakspenger.vedtak.auth.AzureTokenProvider
+import no.nav.tiltakspenger.vedtak.clients.brevpublisher.BrevPublisherGatewayImpl
+import no.nav.tiltakspenger.vedtak.clients.meldekort.MeldekortGrunnlagGatewayImpl
 import no.nav.tiltakspenger.vedtak.clients.utbetaling.UtbetalingClient
+import no.nav.tiltakspenger.vedtak.clients.utbetaling.UtbetalingGatewayImpl
 import no.nav.tiltakspenger.vedtak.db.flywayMigrate
 import no.nav.tiltakspenger.vedtak.repository.InnsendingRepositoryBuilder
 import no.nav.tiltakspenger.vedtak.repository.attestering.AttesteringRepoImpl
 import no.nav.tiltakspenger.vedtak.repository.behandling.PostgresBehandlingRepo
 import no.nav.tiltakspenger.vedtak.repository.behandling.SaksopplysningRepo
 import no.nav.tiltakspenger.vedtak.repository.behandling.VurderingRepo
+import no.nav.tiltakspenger.vedtak.repository.multi.MultiRepoImpl
 import no.nav.tiltakspenger.vedtak.repository.sak.PostgresPersonopplysningerRepo
 import no.nav.tiltakspenger.vedtak.repository.sak.PostgresSakRepo
-import no.nav.tiltakspenger.vedtak.repository.søker.SøkerRepository
+import no.nav.tiltakspenger.vedtak.repository.søker.SøkerRepositoryImpl
 import no.nav.tiltakspenger.vedtak.repository.vedtak.VedtakRepoImpl
 import no.nav.tiltakspenger.vedtak.routes.vedtakApi
 import no.nav.tiltakspenger.vedtak.service.behandling.BehandlingServiceImpl
@@ -48,7 +52,6 @@ internal class ApplicationBuilder(@Suppress("UNUSED_PARAMETER") config: Map<Stri
                 innsendingMediator = innsendingMediator,
                 søkerMediator = søkerMediator,
                 innsendingAdminService = innsendingAdminService,
-                personopplysningService = personopplysningServiceImpl,
                 attesteringRepo = attesteringRepo,
             )
         }
@@ -60,19 +63,30 @@ internal class ApplicationBuilder(@Suppress("UNUSED_PARAMETER") config: Map<Stri
 
     private val sakRepo = PostgresSakRepo()
     private val utbetalingClient = UtbetalingClient(getToken = tokenProviderUtbetaling::getToken)
-    private val utbetalingService = UtbetalingServiceImpl(utbetalingClient, sakRepo)
-    private val søkerRepository = SøkerRepository()
+    private val utbetalingGateway = UtbetalingGatewayImpl(utbetalingClient)
+    private val brevPublisherGateway = BrevPublisherGatewayImpl(rapidsConnection)
+    private val meldekortGrunnlagGateway = MeldekortGrunnlagGatewayImpl(rapidsConnection)
+    private val utbetalingService = UtbetalingServiceImpl(utbetalingGateway, sakRepo)
+    private val søkerRepository = SøkerRepositoryImpl()
     private val behandlingRepo = PostgresBehandlingRepo()
     private val saksopplysningRepo = SaksopplysningRepo()
     private val vurderingRepo = VurderingRepo()
-    private val personopplysningRepo = PostgresPersonopplysningerRepo()
     private val attesteringRepo = AttesteringRepoImpl()
     private val vedtakRepo = VedtakRepoImpl(behandlingRepo, saksopplysningRepo, vurderingRepo)
-    private val vedtakService = VedtakServiceImpl(utbetalingService, vedtakRepo, personopplysningRepo, rapidsConnection)
+    private val multiRepo = MultiRepoImpl(behandlingRepo, attesteringRepo, vedtakRepo)
+    private val personopplysningRepo = PostgresPersonopplysningerRepo()
+    private val vedtakService = VedtakServiceImpl(vedtakRepo)
     private val søkerService = SøkerServiceImpl(søkerRepository)
     private val personopplysningServiceImpl = PersonopplysningServiceImpl(personopplysningRepo)
     private val behandlingService =
-        BehandlingServiceImpl(behandlingRepo, vedtakService, attesteringRepo, personopplysningServiceImpl)
+        BehandlingServiceImpl(
+            behandlingRepo,
+            personopplysningRepo,
+            utbetalingService,
+            brevPublisherGateway,
+            meldekortGrunnlagGateway,
+            multiRepo,
+        )
     private val sakService =
         SakServiceImpl(sakRepo = sakRepo, behandlingRepo = behandlingRepo, behandlingService = behandlingService)
