+++ app/src/main/kotlin/no/nav/tiltakspenger/vedtak/clients/meldekort/MeldekortGrunnlagDTO.kt
@@ -1,0 +1,50 @@
+package no.nav.tiltakspenger.vedtak.clients.meldekort
+
+import java.time.LocalDate
+
+data class MeldekortGrunnlagDTO(
+    val vedtakId: String,
+    val sakId: String,
+    val behandlingId: String,
+    val status: StatusDTO,
+    val vurderingsperiode: PeriodeDTO,
+    val tiltak: List<TiltakDTO>,
+    val personopplysninger: PersonopplysningerDTO,
+    val utfallsperioder: List<UtfallsperiodeDTO>,
+)
+
+data class UtfallsperiodeDTO(
+    val fom: LocalDate,
+    val tom: LocalDate,
+    val antallBarn: Int,
+    val utfall: UtfallForPeriodeDTO,
+)
+
+enum class UtfallForPeriodeDTO {
+    GIR_RETT_TILTAKSPENGER,
+    GIR_IKKE_RETT_TILTAKSPENGER,
+    KREVER_MANUELL_VURDERING,
+}
+
+data class PersonopplysningerDTO(
+    val fornavn: String,
+    val etternavn: String,
+    val ident: String,
+)
+
+enum class StatusDTO {
+    AKTIV,
+    IKKE_AKTIV,
+}
+
+data class TiltakDTO(
+    val periodeDTO: PeriodeDTO,
+    val typeBeskrivelse: String,
+    val typeKode: String,
+    val antDagerIUken: Float,
+)
+
+data class PeriodeDTO(
+    val fra: LocalDate,
+    val til: LocalDate,
+)
