+++ app/src/main/kotlin/no/nav/tiltakspenger/vedtak/clients/meldekort/MeldekortGrunnlagGatewayImpl.kt
@@ -1,0 +1,27 @@
+package no.nav.tiltakspenger.vedtak.clients.meldekort
+
+import mu.KotlinLogging
+import no.nav.helse.rapids_rivers.JsonMessage
+import no.nav.helse.rapids_rivers.RapidsConnection
+import no.nav.tiltakspenger.domene.vedtak.Vedtak
+import no.nav.tiltakspenger.vedtak.service.ports.MeldekortGrunnlagGateway
+import java.time.LocalDateTime
+
+private val LOG = KotlinLogging.logger {}
+private val SECURELOG = KotlinLogging.logger("tjenestekall")
+
+class MeldekortGrunnlagGatewayImpl(
+    private val rapidsConnection: RapidsConnection,
+) : MeldekortGrunnlagGateway {
+    override fun sendMeldekortGrunnlag(vedtak: Vedtak) {
+        mutableMapOf(
+            "@event_name" to "meldekortGrunnlag",
+            "@opprettet" to LocalDateTime.now(),
+            "meldekortGrunnlag" to MeldekortGrunnlagDTOMapper.mapMeldekortGrunnlagDTO(vedtak),
+        ).let { JsonMessage.newMessage(it) }
+            .also { message ->
+                SECURELOG.info { "Vi sender grunnlag : ${message.toJson()}" }
+                rapidsConnection.publish(vedtak.id.toString(), message.toJson())
+            }
+    }
+}
