+++ app/src/main/kotlin/no/nav/tiltakspenger/vedtak/service/personopplysning/PersonopplysningServiceImpl.kt
@@ -2,7 +2,7 @@ package no.nav.tiltakspenger.vedtak.service.personopplysning
 
 import no.nav.tiltakspenger.domene.personopplysninger.SakPersonopplysninger
 import no.nav.tiltakspenger.felles.SakId
-import no.nav.tiltakspenger.vedtak.repository.sak.PersonopplysningerRepo
+import no.nav.tiltakspenger.vedtak.service.ports.PersonopplysningerRepo
 
 class PersonopplysningServiceImpl(
     private val personopplysningerRepo: PersonopplysningerRepo,
