+++ app/src/main/kotlin/no/nav/tiltakspenger/vedtak/service/ports/BrevPublisherGateway.kt
@@ -1,0 +1,8 @@
+package no.nav.tiltakspenger.vedtak.service.ports
+
+import no.nav.tiltakspenger.domene.personopplysninger.PersonopplysningerSøker
+import no.nav.tiltakspenger.domene.vedtak.Vedtak
+
+interface BrevPublisherGateway {
+    fun sendBrev(vedtak: Vedtak, personopplysninger: PersonopplysningerSøker)
+}
