+++ app/src/main/kotlin/no/nav/tiltakspenger/vedtak/service/ports/MeldekortGrunnlagGateway.kt
@@ -1,0 +1,7 @@
+package no.nav.tiltakspenger.vedtak.service.ports
+
+import no.nav.tiltakspenger.domene.vedtak.Vedtak
+
+interface MeldekortGrunnlagGateway {
+    fun sendMeldekortGrunnlag(vedtak: Vedtak)
+}
