+++ app/src/main/kotlin/no/nav/tiltakspenger/vedtak/service/ports/PersonopplysningerRepo.kt
@@ -1,4 +1,4 @@
-package no.nav.tiltakspenger.vedtak.repository.sak
+package no.nav.tiltakspenger.vedtak.service.ports
 
 import no.nav.tiltakspenger.domene.personopplysninger.SakPersonopplysninger
 import no.nav.tiltakspenger.felles.SakId
