+++ app/src/main/kotlin/no/nav/tiltakspenger/vedtak/service/vedtak/VedtakService.kt
@@ -1,7 +1,5 @@
 package no.nav.tiltakspenger.vedtak.service.vedtak
 
-import kotliquery.TransactionalSession
-import no.nav.tiltakspenger.domene.behandling.BehandlingIverksatt
 import no.nav.tiltakspenger.domene.vedtak.Vedtak
 import no.nav.tiltakspenger.felles.BehandlingId
 import no.nav.tiltakspenger.felles.VedtakId
@@ -9,5 +7,4 @@ import no.nav.tiltakspenger.felles.VedtakId
 interface VedtakService {
     fun hentVedtak(vedtakId: VedtakId): Vedtak?
     fun hentVedtakForBehandling(behandlingId: BehandlingId): List<Vedtak>
-    suspend fun lagVedtakForBehandling(behandling: BehandlingIverksatt, tx: TransactionalSession): Vedtak
 }
