+++ app/src/test/kotlin/no/nav/tiltakspenger/vedtak/repository/sak/SakRepoTest.kt
@@ -10,6 +10,7 @@ import no.nav.tiltakspenger.felles.januar
 import no.nav.tiltakspenger.felles.mars
 import no.nav.tiltakspenger.vedtak.db.PostgresTestcontainer
 import no.nav.tiltakspenger.vedtak.db.flywayCleanAndMigrate
+import no.nav.tiltakspenger.vedtak.service.ports.SakRepo
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.testcontainers.junit.jupiter.Container
